@@ -1,44 +1,52 @@
 {
-  "name": "touchstonejs",
-  "version": "0.1.2",
-  "description": "A Mobile App Framework built with ReactJS",
-  "main": "index.js",
-  "author": "Jed Watson",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/JedWatson/touchstonejs.git"
-  },
-  "bugs": {
-    "url": "https://github.com/JedWatson/touchstonejs/issues"
-  },
-  "dependencies": {
-    "classnames": "^1.0.0",
-    "react": "^0.12.2",
-    "react-tappable": "^0.2.3",
-    "underscore": "^1.7.0"
-  },
-  "devDependencies": {
-    "del": "^1.1.0",
-    "gulp": "^3.8.10",
-    "gulp-bump": "^0.1.11",
-    "gulp-connect": "^2.2.0",
-    "gulp-gh-pages": "^0.4.0",
-    "gulp-git": "^0.5.5",
-    "gulp-less": "^3.0.1"
-  },
-  "browserify": {
-    "transform": [
-      "reactify"
-    ]
-  },
-  "keywords": [
-    "react",
-    "react-component",
-    "ui",
-    "mobile",
-    "hybrid"
-  ],
-  "scripts": {
-    "test": "echo \"no tests yet\" && exit 0"
-  }
+	"name": "touchstonejs",
+	"version": "0.3.2",
+	"description": "A Mobile App Framework built with ReactJS",
+	"main": "index.js",
+	"author": "Jed Watson",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/JedWatson/touchstonejs.git"
+	},
+	"bugs": {
+		"url": "https://github.com/JedWatson/touchstonejs/issues"
+	},
+	"dependencies": {
+		"babel-core": "^5.4.7",
+		"babel-plugin-object-assign": "^1.1.0",
+		"blacklist": "^1.1.2",
+		"classnames": "^2.1.2",
+		"gulp-babel": "^5.1.0",
+		"react": "^0.13.3",
+		"react-tappable": "^0.5.0-beta",
+		"xtend": "^4.0.0"
+	},
+	"devDependencies": {
+		"del": "^1.2.0",
+		"gulp": "^3.8.11",
+		"gulp-bump": "^0.3.1",
+		"gulp-connect": "^2.2.0",
+		"gulp-gh-pages": "^0.5.1",
+		"gulp-git": "^1.2.4",
+		"gulp-less": "^3.0.3",
+		"happiness": "^1.0.5"
+	},
+	"keywords": [
+		"react",
+		"react-component",
+		"ui",
+		"mobile",
+		"hybrid"
+	],
+	"scripts": {
+		"build": "gulp build:lib",
+		"watch": "gulp watch:lib",
+		"lint": "happiness",
+		"test": "npm run lint"
+	},
+	"standard": {
+		"ignore": [
+			"**/lib/**"
+		]
+	}
 }
