@@ -1,26 +1,53 @@
-var del = require('del'),
-	gulp = require('gulp'),
-	bump = require('gulp-bump'),
-	connect = require('gulp-connect'),
-	deploy = require("gulp-gh-pages"),
-	git = require("gulp-git"),
-	less = require('gulp-less');
+var babel = require('gulp-babel');
+var bump = require('gulp-bump');
+var connect = require('gulp-connect');
+var del = require('del');
+var deploy = require('gulp-gh-pages');
+var git = require('gulp-git');
+var gulp = require('gulp');
+var less = require('gulp-less');
 
-
+var src = [
+	'src/**/*.js',
+	'!**/__tests__/**/*'
+];
 /**
  * Clean the build
  */
 
-gulp.task('clean', function(done) {
-	del(['site/public/build'], done);
+gulp.task('clean:lib', function (done) {
+	del(['lib'], done);
 });
 
+/**
+ * Build lib
+ */
+
+gulp.task('build:lib', function () {
+	return gulp.src(src)
+		.pipe(babel({
+			plugins: [require('babel-plugin-object-assign')]
+		}))
+		.pipe(gulp.dest('lib'));
+});
+
+gulp.task('watch:lib', ['build:lib'], function () {
+	gulp.watch(src, ['build:lib']);
+});
+
+/**
+ * Clean site build
+ */
+
+gulp.task('clean:site', function (done) {
+	del(['site/public/build'], done);
+});
 
 /**
  * Build site css from less
  */
 
-function buildSiteCSS() {
+function buildSiteCSS () {
 	return gulp.src('site/src/less/site.less')
 		.pipe(less())
 		.pipe(gulp.dest('site/public/build/css'))
@@ -28,7 +55,6 @@ function buildSiteCSS() {
 }
 
 gulp.task('build:site:css', buildSiteCSS);
-
 
 /**
  * Build site
@@ -40,16 +66,15 @@ gulp.task('build:site', [
 
 gulp.task('watch:site', [
 	'build:site:css'
-], function() {
+], function () {
 	gulp.watch(['site/src/less/**/*'], ['build:site:css']);
 });
-
 
 /**
  * Serve task for local development
  */
 
-gulp.task('site', ['watch:site'], function() {
+gulp.task('site', ['watch:site'], function () {
 	connect.server({
 		root: 'site/public',
 		port: 8000,
@@ -57,15 +82,12 @@ gulp.task('site', ['watch:site'], function() {
 	});
 });
 
-
-
-
 /**
  * Version bump tasks
  */
 
-function _bump(type) {
-	return function() {
+function _bump (type) {
+	return function () {
 		return gulp.src(['./package.json', './bower.json'])
 			.pipe(bump({ type: type }))
 			.pipe(gulp.dest('./'));
@@ -76,13 +98,12 @@ gulp.task('bump', _bump('patch'));
 gulp.task('bump:minor', _bump('minor'));
 gulp.task('bump:major', _bump('major'));
 
-
 /**
  * Git tag task
  * (version *must* be bumped first)
  */
 
-gulp.task('publish:tag', function(done) {
+gulp.task('publish:tag', function (done) {
 	var pkg = require('./package.json');
 	var v = 'v' + pkg.version;
 	var message = 'Release ' + v;
@@ -96,24 +117,22 @@ gulp.task('publish:tag', function(done) {
 	});
 });
 
-
 /**
  * npm publish task
  * * (version *must* be bumped first)
  */
 
-gulp.task('publish:npm', function(done) {
+gulp.task('publish:npm', function (done) {
 	require('child_process')
 		.spawn('npm', ['publish'], { stdio: 'inherit' })
 		.on('close', done);
 });
 
-
 /**
  * Deploy tasks
  */
 
-gulp.task('publish:site', ['build:site'], function() {
+gulp.task('publish:site', ['build:site'], function () {
 	return gulp.src('site/public/**/*').pipe(deploy());
 });
 
