@@ -14,7 +14,7 @@ module.exports = React.createClass({
 		var className = classnames({
 			'item-media':   true,
 			'is-icon':      this.props.icon,
-			'is-avatar':    this.props.avatar,
+			'is-avatar':    this.props.avatar || this.props.avatarInitials,
 			'is-thumbnail': this.props.thumbnail
 		}, this.props.className);
 
@@ -22,9 +22,9 @@ module.exports = React.createClass({
 		var icon = this.props.icon ? (
 			<div className={'item-icon ' + this.props.icon} />
 		) : null;
-		var avatar = this.props.avatar ? (
+		var avatar = this.props.avatar || this.props.avatarInitials ? (
 			<div className="item-avatar">
-				<img src={this.props.avatar} />
+				{this.props.avatar ? <img src={this.props.avatar} /> : this.props.avatarInitials}
 			</div>
 		) : null;
 		var thumbnail = this.props.thumbnail ? (
