@@ -1,6 +1,7 @@
-var React = require('react/addons');
+var blacklist = require('blacklist')
 var	classnames = require('classnames');
-var _ = require('underscore');
+
+var React = require('react/addons');
 
 module.exports = React.createClass({
 	displayName: 'LabelTextarea',
@@ -10,28 +11,20 @@ module.exports = React.createClass({
 		};
 	},
 	render: function() {
-		var disabled = this.props.disabled || this.props.readonly;
 		var className = classnames(this.props.className, {
 			'list-item': true,
 			'field-item': true,
 			'align-top': true,
 			'is-first': this.props.first,
-			'u-selectable': disabled
+			'u-selectable': this.props.disabled
 		});
 
-		// FIXME: we can do better
-		var curated = _.extend({}, this.props);
-		delete curated.className;
-		delete curated.disabled;
-		delete curated.first;
-		delete curated.readonly;
-		delete curated.children;
-		delete curated.label;
+		var props = blacklist(this.props, 'children', 'className', 'disabled', 'first', 'label', 'readonly')
 
 		var renderInput = this.props.readonly ? (
 			<div className="field u-selectable">{this.props.value}</div>
 		) : (
-			<textarea disabled={disabled} {...curated} className="field" />
+			<textarea {...props} className="field" />
 		);
 
 		return (
