@@ -1,6 +1,4 @@
-/** @jsx React.DOM */
-
-var React    = require('react');
+var React = require('react');
 var Tappable = require('react-tappable');
 
 module.exports = React.createClass({
